Adds example code for signing and validation
An examples folder with a signer and a validator application
is added. These are based on gStreamer and requires
additional dependencies to build.

The example applications are not complete, but serves so far
as example code.

A shell script to build the applications is added.
Further, to be able to build the library and the applications,
meson files have been added and also a dummy definitions file,
which will be replaced when the proper code has been added.

Also, workflow actions are added for testing in the repository.