Enables crypto_handle in session struct
In addition, changed the onvif_media_signing_product_info_t to
have a static size memory to become more clear to the user about
the constraints when adding them to the SEI.
The example apps have been updated according to that.

The README has been updated with the examples folder in the tree.