Adds NAL Unit to prepend and num_pending_seis to get sei API
Now the user can get help when fetching SEIs. If SEIs are added to
the session at once (in place) it has to follow the standard and
SEIs should be part of other headers of a primary slice.
Also, it is possible to get information on how many SEIs there are
in the pipe, which can help the user to wait for ongoing signing.