Adds signer test for golden SEI (#54)
The test generates and adds a golden SEI at the beginning of the
test stream. The oms_setting struct has been extended with a
with_golden_sei member. The verify_sei is updated with checks
on golden SEI.

A helper function to determine if a TLV has any mandatory tag
or not has been implemented.

Co-authored-by: bjornvolcker <bjornvolcker@users.noreply.github.com>