Adds support for validating lost P-frames
Validating a video with lost P-frames is now supported. When
verifying the gop hash fails, individual hashes are checked, if
a hash list exists in the SEI that is.
Tests for normal operation as well as signig partial and multiple
gops have been added.

In addition, the multiple gop test now is set to sign every third
GOP, which better exercise the recursive parts of the code.