Extracts public key from certificate (#98)
Up till now, the certificate has been used as the public key. With
this commit, private keys and certificates for both EC and RSA have
been pre-generated for the tests. A helper function to read these
files is present, which also is used by the signer application.
The test certificate is self signed at this point.

Co-authored-by: bjornvolcker <bjornvolcker@users.noreply.github.com>