Adds support for certificate verification
The tests currently use a self signed certificate, which is like
using a trusted CA directly. Tests and fixes for true certificate
chains will be handled in a different commit.

Proper provenance is now set and an extra level where a certificate
chain could be verified without a trusted certificate has been
added.