Unsigned GOPs are no longer reported
Since it can only be known after processing the entire file,
reporting 'unsigned' video will not be done in intermediate
authenticity reports. The user is recommended to look in the
accumulated report for that information.

As a concequence of this change also 'stream is signed' will
only be reported once, that is, on the first infeasible SEI.
This affects scrubbing, when it is no longer reported after
a scrub. The user already knows that the video is signed.

Tests have been updated.