Adds a secondary MD context
The signing side will need two EVP_MD_CTX, one for updating
the NAL Unit hash when the NAL Unit is split in parts, and
one for updating the GOP hash. The latter is a hash of the
hash list, but there is no guarantee that all hashes will
fit in a list.

This commit extends the crypto handle with a second CTX,
but everything else is untouched.