Renames public API onvif_media_signing_set_max_signing_nalus
The new name is onvif_media_signing_set_max_signing_frames()
and instead of counting NAL Units, the library counts frames.
A frame can be split in several NAL Units which makes it
more difficult to configure by the user.
Also it avoids ending up hashing one NAL unit of a frame and
then signing. This would then put half of the frame in one
SEI and the other half in the next one.