Adds a test for validating a file with two useless seis
When exporting a signed stream to a file at least one SEI in the
beginning will not be possible to use. The reason is that it
is associated with a GOP prior to the exported snippet. The
validation should not be dependent on assuming only one SEI to
be useless. It should be possible to handle any number of SEIs
that are associated with GOPs prior to the export. This can for
example happen when the signing device (e.g., SE/TPM/TEE) was
blocked just when export started.

Without any changes to the test it would have failed, so before
a solution for synchronizing the first SEI is in place the test
has been adjusted to pass (wrongly).