Adds update_sei_in_validation
At the beginning of a session when SEIs are not in sync they
will be used in attempts for validation. Therefore, a SEI that
is in_validation needs to get its validation_status reset
between these attempts.

A general function that applies operations on the SEI
in_validation is added and called in some places.

This commit affects four tests. These tests have been changed
temporarily until the full implementation has been completed.