Fixes bugs in signer example application
This commit fixes two bugs present in the signer example app.
The first bug added SEIs in the wrong order if multiple ones
were fetched at the same time.
The second bug was that multiple NAL Units in an AU were added
for signing as if there was only one. The root cause is that
the GStreamer muxer elements can only handle AU alignments of
the data stream. To fix this, the memory is parsed for multiple
NAL Units before adding for signing.

Further, minor editorial changes were also made.