Changes version numbering (#219)
The version now is the same as specification revision.
So if the specification revision changes from 25.12 to
26.6 the version should be changed to r26.6.0.

Three tags have already been released with the old way.
These are converted when parsed to strings. They get
version r24.12.x, since they were released on top of
specification 24.12.

In addition, makes sure to print version on signing side
if validation runs in batch mode.

Bumps version to r25.12.0.